Refactor share target platform implementation
