+++ js/app/tauri/src-tauri/src/share_target/noop.rs
@@ -1,0 +1,40 @@
+use super::{
+    PendingShareFilesState, ShareTargetPlatform, StagedSharedFile, lock_pending_share_filenames,
+};
+use tauri::AppHandle;
+use url::Url;
+
+pub(super) struct ShareTargetPlatformImpl;
+
+impl ShareTargetPlatform for ShareTargetPlatformImpl {
+    fn cleanup_stale_staged_shared_files(_app: &AppHandle) {}
+
+    fn get_pending_share_filenames(_app: AppHandle, state: &PendingShareFilesState) -> Vec<String> {
+        lock_pending_share_filenames(state).clone()
+    }
+
+    fn pop_shared_files(
+        _app: AppHandle,
+        _filenames: Vec<String>,
+        _state: &PendingShareFilesState,
+    ) -> Vec<StagedSharedFile> {
+        vec![]
+    }
+
+    fn clear_shared_files(_app: AppHandle, _tokens: Vec<String>) -> Result<(), String> {
+        Ok(())
+    }
+
+    async fn upload_shared_file_to_presigned_url(
+        _app: AppHandle,
+        _token: String,
+        _upload_url: String,
+        _mime_type: String,
+    ) -> Result<(), String> {
+        Ok(())
+    }
+
+    fn maybe_handle_share_deep_link(_handle: &AppHandle, _url: &Url) -> bool {
+        false
+    }
+}
